Move the default fetch handler to a seperate module
