+++ lib/runRouter.js
@@ -1,8 +1,8 @@
 'use strict';
 
 var Router = require('react-router'),
-    join   = require('bluebird').join,
-    contains = require('ramda').contains
+    contains = require('ramda').contains,
+    defaultFetchHandler = require('./defaultFetchHandler')
 
 /**
  * Run react-router then check if the matching routes need to fetch some data
@@ -13,15 +13,14 @@ var Router = require('react-router'),
  * @param {function} cb
  */
 module.exports = function runRouter(routes, routerArg, fetchHandler, cb) {
-    var components = [],
-        fetches    = []
+    fetchHandler = typeof fetchHandler === 'function' ? fetchHandler : defaultFetchHandler
 
-    var addFetch = function addFetch(component) {
+    var components = []
+
+    var addComponent = function addComponent(component) {
         if (typeof component.fetchData === 'function') {
-            var name = component.displayName || component.name
-            if (!contains(name)(components)) {
-                components.push(name)
-                fetches.push(component.fetchData())
+            if (!contains(component)(components)) {
+                components.push(component)
             }
         }
     }
@@ -29,7 +28,7 @@ module.exports = function runRouter(routes, routerArg, fetchHandler, cb) {
     var parseDependencies = function parseDependencies(component) {
         if (typeof component.getDependencies === 'function') {
             component.getDependencies().forEach(function(dep) {
-                addFetch(dep)
+                addComponent(dep)
             })
         }
     }
@@ -39,37 +38,16 @@ module.exports = function runRouter(routes, routerArg, fetchHandler, cb) {
 
         state.routes.forEach(function(matchedRoute) {
             parseDependencies(matchedRoute.handler)
-            addFetch(matchedRoute.handler)
+            addComponent(matchedRoute.handler)
         })
 
-        if (fetches.length) {
-            if (typeof fetchHandler === 'function') {
-                fetchHandler(components).then(function(data) {
-                    cb(Handler, data)
-                })
-            }
-            else {
-                fetches.push(function() {
-                    var args = Array.prototype.slice.call(arguments),
-                        data = {},
-                        i = 0
-
-                    components.forEach(function(component) {
-                        data[component] = args[i]
-                        i++
-                    })
-
-                    components = []
-                    fetches    = []
-                    cb(Handler, data)
-                })
-
-                join.apply(null, fetches)
-            }
+        if (components.length) {
+            fetchHandler(components).then(function(data) {
+                cb(Handler, data)
+            })
         }
         else {
             components = []
-            fetches    = []
             cb(Handler)
         }
     })
